Admin Profile, Online Orders, Subscription Success, WaterOrders and S…

+++ Model/superAdminClients.cs
@@ -18,7 +18,7 @@ namespace WRS2big_Web.Model
 
         public DateTime dateSubscribed { get; set; }
         public string status { get; set; }
-        public int amount { get; set; }
+        public decimal amount { get; set; }
         public DateTime subExpiration { get; set; }
         public string currentSubStatus { get; set; }
         public string paymentStatus { get; set; }
